♻️ Refactor LikeCoinButton code
